+++ src/scholarag/app/routers/qa.py
@@ -104,7 +104,7 @@ async def passthrough(
     -------
         A single answer.
     """  # noqa: D301, D400, D205
-    if len(request.query) > 10000:
+    if len(request.query) > settings.misc.query_max_size:
         raise HTTPException(status_code=413, detail="Query string too long.")
     start = time.time()
     logger.info("Finding answers ...")
@@ -242,7 +242,7 @@ async def generative_qa(
     -------
         A single answer with metadata of each relevant source.
     """  # noqa: D301, D400, D205
-    if len(request.query) > 10000:
+    if len(request.query) > settings.misc.query_max_size:
         raise HTTPException(status_code=413, detail="Query string too long.")
     start = time.time()
     logger.info("Finding answers ...")
@@ -464,7 +464,7 @@ async def streamed_generative_qa(
     -------
         A single answer with metadata of each relevant source.
     """  # noqa: D301, D400, D205
-    if len(request.query) > 10000:
+    if len(request.query) > settings.misc.query_max_size:
         raise HTTPException(status_code=413, detail="Query string too long.")
     start = time.time()
     logger.info("Finding answers ...")